LayoutTests/imported/w3c/web-platform-tests/XMLHttpRequest/send-network-error-sync-events.sub-expected.txt

CONSOLE MESSAGE: line 24: XMLHttpRequest cannot load http://nonexistent-origin.localhost}:8800. Cross origin requests are only supported for HTTP.

FAIL XmlHttpRequest: The send() method: Throw a "throw an "NetworkError" exception when Network error happens (synchronous flag is set) assert_throws: function "function ()
            {
                xhr.send("Test ..." threw object "Error: NETWORK_ERR: XMLHttpRequest Exception 101" that is not a DOMException NetworkError: property "code" is equal to 101, expected 19
